# optimize: skip reading whole column if it's been preread