--- a/pqarrow/arrow.go
+++ b/pqarrow/arrow.go
@@ -858,7 +858,6 @@ func (c *ParquetConverter) writeColumnToArray(
 					}
 					return true
 				})
-
 			} else {
 				w.Write(c.scratchValues)
 			}